Split contract-specific namespace module code into its own file.
